Edit file "combine.sh" and "game.sh"

+++ Gor/19_05_2021/combine.sh
@@ -1,9 +1,18 @@
 #!/bin/bash
 
+if [ -d "math" ]
+then
+	rm -rf math
+fi
+if [ -d "text" ]
+then
+	rm -rf text
+fi
+
 for (( i=1; i<4; i++ ))
 do
 	> math_$i
-	> test_$i
+	> text_$i
 done
 
 for file in ./*
@@ -11,12 +20,18 @@ do
         folder_1=${file##*/}
         folder_2=${folder_1%.*}
         folder_3=${folder_2%_*}
-        if [ -d $folder_3 ]
-        then
-                mv "$file" "./$folder_3"
-        else
-                mkdir $folder_3
-                mv "$file" "./$folder_3"
-        fi
+        if [ $folder_3 = "math" ] || [ $folder_3 = "text" ] 
+	then
+		if [ -d $folder_3 ] && [ -f $file ]
+        	then
+        	        mv "$file" "./$folder_3"
+        	else
+        	        mkdir $folder_3
+        	        mv "$file" "./$folder_3"
+        	fi
+	fi
 done
-
+echo 'ls directory math'$'\n'
+ls math
+echo $'\n''ls directory text'$'\n'
+ls text
